requirements.txt
flask
python-docx
requests
PyMuPDF
twilio
chatterbot
pytesseract
Pillow